+++ src/clients/server.generated.ts
@@ -6403,7 +6403,7 @@ or a different price that is not a discount */
     /** Optional discount amount */
     discount!: number;
     /** Any comments regarding this product instance */
-    comments?: string;
+    details?: string;
 
     constructor(data?: IProductInstance) {
         if (data) {
@@ -6439,7 +6439,7 @@ or a different price that is not a discount */
             }
             this.basePrice = _data["basePrice"];
             this.discount = _data["discount"];
-            this.comments = _data["comments"];
+            this.details = _data["details"];
         }
     }
 
@@ -6470,7 +6470,7 @@ or a different price that is not a discount */
         }
         data["basePrice"] = this.basePrice;
         data["discount"] = this.discount;
-        data["comments"] = this.comments;
+        data["details"] = this.details;
         return data; 
     }
 }
@@ -6503,7 +6503,7 @@ or a different price that is not a discount */
     /** Optional discount amount */
     discount: number;
     /** Any comments regarding this product instance */
-    comments?: string;
+    details?: string;
 }
 
 export class Company implements ICompany {
@@ -6886,6 +6886,10 @@ export enum InvoiceStatus {
 export enum ActivityType {
     STATUS = "STATUS",
     COMMENT = "COMMENT",
+    EDIT = "EDIT",
+    REASSIGN = "REASSIGN",
+    ADDPRODUCT = "ADDPRODUCT",
+    DELPRODUCT = "DELPRODUCT",
 }
 
 export class InvoiceActivity implements IInvoiceActivity {
@@ -9027,6 +9031,7 @@ export interface ICompanyListResponse {
 export class CompanySummary implements ICompanySummary {
     id!: number;
     name!: string;
+    logoFilename!: string;
 
     constructor(data?: ICompanySummary) {
         if (data) {
@@ -9041,6 +9046,7 @@ export class CompanySummary implements ICompanySummary {
         if (_data) {
             this.id = _data["id"];
             this.name = _data["name"];
+            this.logoFilename = _data["logoFilename"];
         }
     }
 
@@ -9055,6 +9061,7 @@ export class CompanySummary implements ICompanySummary {
         data = typeof data === 'object' ? data : {};
         data["id"] = this.id;
         data["name"] = this.name;
+        data["logoFilename"] = this.logoFilename;
         return data; 
     }
 }
@@ -9062,6 +9069,7 @@ export class CompanySummary implements ICompanySummary {
 export interface ICompanySummary {
     id: number;
     name: string;
+    logoFilename: string;
 }
 
 export class CompanyParams implements ICompanyParams {
@@ -9423,6 +9431,7 @@ export interface IContractListResponse {
 export class ContractSummary implements IContractSummary {
     id!: number;
     title!: string;
+    value!: number;
     status!: ContractStatus;
 
     constructor(data?: IContractSummary) {
@@ -9438,6 +9447,7 @@ export class ContractSummary implements IContractSummary {
         if (_data) {
             this.id = _data["id"];
             this.title = _data["title"];
+            this.value = _data["value"];
             this.status = _data["status"];
         }
     }
@@ -9453,6 +9463,7 @@ export class ContractSummary implements IContractSummary {
         data = typeof data === 'object' ? data : {};
         data["id"] = this.id;
         data["title"] = this.title;
+        data["value"] = this.value;
         data["status"] = this.status;
         return data; 
     }
@@ -9461,6 +9472,7 @@ export class ContractSummary implements IContractSummary {
 export interface IContractSummary {
     id: number;
     title: string;
+    value: number;
     status: ContractStatus;
 }
 
@@ -9962,6 +9974,7 @@ export class InvoiceSummary implements IInvoiceSummary {
     id!: number;
     title!: string;
     companyId!: number;
+    value!: number;
     status!: InvoiceStatus;
 
     constructor(data?: IInvoiceSummary) {
@@ -9978,6 +9991,7 @@ export class InvoiceSummary implements IInvoiceSummary {
             this.id = _data["id"];
             this.title = _data["title"];
             this.companyId = _data["companyId"];
+            this.value = _data["value"];
             this.status = _data["status"];
         }
     }
@@ -9994,6 +10008,7 @@ export class InvoiceSummary implements IInvoiceSummary {
         data["id"] = this.id;
         data["title"] = this.title;
         data["companyId"] = this.companyId;
+        data["value"] = this.value;
         data["status"] = this.status;
         return data; 
     }
@@ -10003,6 +10018,7 @@ export interface IInvoiceSummary {
     id: number;
     title: string;
     companyId: number;
+    value: number;
     status: InvoiceStatus;
 }
 
@@ -10777,6 +10793,7 @@ export class UserSummary implements IUserSummary {
     lastNamePreposition!: string;
     lastName!: string;
     email!: string;
+    avatarFilename!: string;
 
     constructor(data?: IUserSummary) {
         if (data) {
@@ -10794,6 +10811,7 @@ export class UserSummary implements IUserSummary {
             this.lastNamePreposition = _data["lastNamePreposition"];
             this.lastName = _data["lastName"];
             this.email = _data["email"];
+            this.avatarFilename = _data["avatarFilename"];
         }
     }
 
@@ -10811,6 +10829,7 @@ export class UserSummary implements IUserSummary {
         data["lastNamePreposition"] = this.lastNamePreposition;
         data["lastName"] = this.lastName;
         data["email"] = this.email;
+        data["avatarFilename"] = this.avatarFilename;
         return data; 
     }
 }
@@ -10821,6 +10840,7 @@ export interface IUserSummary {
     lastNamePreposition: string;
     lastName: string;
     email: string;
+    avatarFilename: string;
 }
 
 /** Make all properties in T optional */