+++ src/components/activities/DocumentStatusProps.tsx
@@ -113,6 +113,7 @@ class DocumentStatusProps extends React.Component<Props, State> {
           <PropsButtons
             editing={editing}
             canDelete={undefined}
+            canSave
             entity={documentType}
             status={this.props.resourceStatus}
             cancel={this.cancel}
@@ -153,4 +154,4 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   ),
 });
 
-export default connect(mapDispatchToProps)(DocumentStatusProps);
+export default connect(null, mapDispatchToProps)(DocumentStatusProps);