+++ src/components/navigation/InvoicesMenu.tsx
@@ -15,13 +15,12 @@ function InvoicesMenu() {
         onMouseLeave={() => changeOpen(false)}
         item
         icon={null}
-        as={NavLink}
-        to="/invoice"
-        text={(
-          <>
+        trigger={(
+          <NavLink to="/invoice">
             <Icon name="money bill alternate outline" />
+            {' '}
             Invoices
-          </>
+          </NavLink>
         ) as any}
       >
         <Dropdown.Menu>