+++ src/components/user/UserSelector.tsx
@@ -10,7 +10,6 @@ interface Props {
   value: number;
   options: UserSummary[];
   onChange: (value: number | number[]) => void;
-  // eslint-disable-next-line react/require-default-props
   hideEmail?: boolean;
   correct?: boolean;
 }
@@ -19,9 +18,9 @@ function UserSelector(props: Props & DropdownProps) {
   const [open, changeOpen] = useState(false);
 
   const {
-    value, onChange, options, hideEmail, correct, ...rest
+    value, onChange, options, hideEmail, correct,
   } = props;
-  const dropdownOptions = props.options.map((x) => ({
+  const dropdownOptions = options.map((x) => ({
     key: x.id,
     text: formatContactName(x.firstName, x.lastNamePreposition, x.lastName),
     description: hideEmail ? undefined : x.email,
@@ -35,10 +34,9 @@ function UserSelector(props: Props & DropdownProps) {
       selection
       fluid
       error={(value < 1 && !open) && correct !== true}
-      {...rest}
       options={dropdownOptions}
       value={props.value}
-      onChange={(e, data) => props.onChange(data.value as any)}
+      onChange={(e, data) => onChange(data.value as any)}
       onOpen={() => changeOpen(true)}
       onClose={() => changeOpen(false)}
     />