+++ src/pages/ContractProductInstanceModal.tsx
@@ -41,7 +41,7 @@ interface Props extends SelfProps {
   removeProductInstance: (contractId: number, id: number) => void;
 }
 
-class ProductInstanceModal extends React.Component<Props> {
+class ContractProductInstanceModal extends React.Component<Props> {
   close = () => {
     const { contractId } = this.props.match.params;
     this.props.fetchContract(parseInt(contractId, 10));
@@ -81,7 +81,7 @@ class ProductInstanceModal extends React.Component<Props> {
       productInstance = {
         id: -1,
         contractId: parseInt(contractId, 10),
-        productId: 0,
+        productId: -1,
         basePrice: 0,
         discount: 0,
         comments: '',
@@ -113,7 +113,7 @@ class ProductInstanceModal extends React.Component<Props> {
     let activities;
     if (!create) {
       activities = [
-        <Segment secondary style={{ margin: '2em 1em 1em' }}>
+        <Segment secondary style={{ margin: '2em 1em 1em' }} key="seg-1">
           <FinancialDocumentProgress
             documentId={productInstance.id}
             parentId={productInstance.contractId}
@@ -122,7 +122,7 @@ class ProductInstanceModal extends React.Component<Props> {
             resourceStatus={status}
           />
         </Segment>,
-        <Segment style={{ margin: '2em 1em 1em' }}>
+        <Segment style={{ margin: '2em 1em 1em' }} key="seg-2">
           <ActivitiesList
             activities={productInstance.activities as GeneralActivity[]}
             componentId={productInstance.id}
@@ -186,4 +186,5 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   ),
 });
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(ProductInstanceModal));
+export default withRouter(connect(mapStateToProps,
+  mapDispatchToProps)(ContractProductInstanceModal));